converted all doubles to floats

+++ src/Marker.cpp
@@ -49,12 +49,12 @@ cv::Point Marker::center() {
 
 cv::Point2f Marker::centerf() {
     return cv::Point2f(
-            (double)this->start_x + ((this->end_x - this->start_x) / 2.0),
-            (double)this->start_y + ((this->end_y - this->start_y) / 2.0)
+            (float)this->start_x + ((this->end_x - this->start_x) / 2.0),
+            (float)this->start_y + ((this->end_y - this->start_y) / 2.0)
     );
 }
 
-double Marker::distance(Marker one, Marker two) {
+float Marker::distance(Marker one, Marker two) {
     cv::Point a = one.center();
     cv::Point b = two.center();
     return sqrt(
@@ -62,11 +62,11 @@ double Marker::distance(Marker one, Marker two) {
     );
 }
 
-double Marker::angle(Marker one, Marker two) {
+float Marker::angle(Marker one, Marker two) {
     cv::Point a = one.center();
     cv::Point b = two.center();
-    double dx = b.x - a.x;
-    double dy = b.y - a.y;
+    float dx = b.x - a.x;
+    float dy = b.y - a.y;
     if(dx == 0) {
         return 0;
     }
